Improved listfile and changed dir structure.
* Improved the top-level lisfile, the most notable change is the use
of double quotes on any expansion of variables that should contain
a path, some commands such ass `include_directories' issue a error
message if the variable is empty, using quotes is more robust, for
instance:

.. code-block:: cmake

   set (acme_EXTRA_INCLUDE_DIRS) # Contains null value.

   # Will not issue a error.
   include_directories (src
                        ${acme_EXTRA_INCLUDE_DIRS})

Since this variable value is null `include_directories' will get `src'
ass a parameter an nothing else. The above line is equal as this one:

   include_directories (src

   # until *N* white spaces.
   )

The problem can be corrected by using double quotes.

.. code-block:: cmake

   set (acme_EXTRA_INCLUDE_DIRS) # Contains null value.

   # Will issue a error.
   include_directories (src
                        "${acme_EXTRA_INCLUDE_DIRS}")

* Fixed style in the top level lisfile.

* Move `include/conf.h.in' to `src/conf.h.in', since the header is not
exported it does not make sense to have it in the `include' directory,
normally that directory is used for header that will be used by third
party applications.